move name tabs to the navbar
