+++ src/layout/NamePage.tsx
@@ -6,34 +6,16 @@ import {
   useLocation,
   useOutletContext,
 } from 'react-router-dom';
-import NameButtonNavbar from 'components/ButtonNavbar';
+
 import styled from 'styled-components';
 import { useSubstrate } from 'layout/hooks';
 import { calcBlockTimeMs } from 'substrate/utils';
-
-const getTabBasePath = (currentPath) => {
-  let currentParts = currentPath.split('/');
-  currentParts.pop();
-  return currentParts.join('/');
-};
-
-const getTabs = (tabTitles: string[], basePath: string) => {
-  let tabs = tabTitles.map((title) => ({
-    title,
-    path: `${basePath}/${title.toLowerCase()}`,
-  }));
-  return tabs;
-};
-
 interface NamePageProps {
   className?: string;
 }
 const NamePage = ({ className }: NamePageProps) => {
   let { name } = useParams();
   let location = useLocation();
-  let tabBasePath = getTabBasePath(location.pathname);
-  let tabTitles = ['Register', 'Details', 'Subdomains'];
-  let tabs = getTabs(tabTitles, tabBasePath);
   const { api, nameServiceProvider }: any = useSubstrate();
   const [nameRegistration, setNameRegistration] = useState();
   useEffect(() => {
@@ -57,7 +39,6 @@ const NamePage = ({ className }: NamePageProps) => {
             <div className="d-flex align-items-center">
               <div className="fw-light fs-4">{name}</div>
             </div>
-            <NameButtonNavbar tabs={tabs} />
           </div>
         </Card.Header>
         <Card.Body>
