+++ src/substrate/utils.ts
@@ -3,7 +3,6 @@ import { cryptoWaitReady } from '@polkadot/util-crypto';
 import { web3FromSource } from '@polkadot/extension-dapp';
 import { BN } from 'bn.js';
 import { bnMin } from '@polkadot/util';
-import { timeStamp } from 'console';
 
 /**
  * trims the specified character from the end of the string.
